requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.0
protobuf==2.6.1
pycrypto==2.6.1
python-axolotl==0.1.7
python-axolotl-curve25519==0.1
python-dateutil==2.4.2
six==1.10.0
vobject==0.6.6
Werkzeug==0.11.3
wheel==0.24.0
yowsup2==2.4.48
